// frontend/src/lib/api/generated/model/complianceDocType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Construction ERP API
 * OpenAPI spec version: 0.1.0
 */

export type ComplianceDocType = typeof ComplianceDocType[keyof typeof ComplianceDocType];


export const ComplianceDocType = {
  tax_clearance: 'tax_clearance',
  vat_registration: 'vat_registration',
  company_registration: 'company_registration',
  public_liability: 'public_liability',
  workmans_compensation: 'workmans_compensation',
  safety_certificate: 'safety_certificate',
  bank_confirmation: 'bank_confirmation',
  trade_licence: 'trade_licence',
  other: 'other',
} as const;